+++ lib/core/inject/intect.dart
@@ -1,23 +1,23 @@
 import 'package:get_it/get_it.dart';
-import 'package:music_app/layers/domain/repositories/get_albums_by_artist_repository.dart';
-import 'package:music_app/layers/domain/repositories/get_all_artists_repository.dart';
-import 'package:music_app/layers/domain/repositories/get_musics_by_artist_repository.dart';
-import 'package:music_app/layers/domain/usecases/get_albums_by_artist/get_albums_by_artist_usecase.dart';
-import 'package:music_app/layers/domain/usecases/get_albums_by_artist/get_albums_by_artist_usecase_imp.dart';
-import 'package:music_app/layers/domain/usecases/get_all_artists/get_all_artists_usecase.dart';
-import 'package:music_app/layers/infra/datasources/get_all_artists_datasource.dart';
-import 'package:music_app/layers/infra/repositories/get_albums_by_artist_repository_imp.dart';
-import 'package:music_app/layers/infra/repositories/get_musics_by_artist_repository_imp.dart';
 
-import '../../layers/domain/usecases/get_all_artists/get_all_artists_usecase_imp.dart';
-import '../../layers/domain/usecases/get_musics_by_artist/get_musics_by_artist_usecase.dart';
-import '../../layers/domain/usecases/get_musics_by_artist/get_musics_by_artist_usecase_imp.dart';
-import '../../layers/external/local/get_albums_by_artist_local_datasource_imp.dart';
-import '../../layers/external/local/get_all_artists_local_datasource_imp.dart';
-import '../../layers/external/local/get_musics_by_artist_local_datasource_imp.dart';
-import '../../layers/infra/datasources/get_albums_by_artist_datasource.dart';
-import '../../layers/infra/datasources/get_musics_by_artists_datasource.dart';
-import '../../layers/infra/repositories/get_all_artists_repository_imp.dart';
+import '../../modules/album/domain/repositories/get_albums_by_artist_repository.dart';
+import '../../modules/album/domain/usecases/get_albums_by_artist_usecase.dart';
+import '../../modules/album/domain/usecases/get_albums_by_artist_usecase_imp.dart';
+import '../../modules/album/external/stub/get_albums_by_artist_local_datasource_imp.dart';
+import '../../modules/album/infra/datasources/get_albums_by_artist_datasource.dart';
+import '../../modules/album/infra/repositories/get_albums_by_artist_repository_imp.dart';
+import '../../modules/artist/domain/repositories/get_all_artists_repository.dart';
+import '../../modules/artist/domain/usecases/get_all_artists_usecase.dart';
+import '../../modules/artist/domain/usecases/get_all_artists_usecase_imp.dart';
+import '../../modules/artist/external/stub/get_all_artists_local_datasource_imp.dart';
+import '../../modules/artist/infra/datasources/get_all_artists_datasource.dart';
+import '../../modules/artist/infra/repositories/get_all_artists_repository_imp.dart';
+import '../../modules/music/domain/repositories/get_musics_by_artist_repository.dart';
+import '../../modules/music/domain/usecases/get_musics_by_artist_usecase.dart';
+import '../../modules/music/domain/usecases/get_musics_by_artist_usecase_imp.dart';
+import '../../modules/music/external/stub/get_musics_by_artist_local_datasource_imp.dart';
+import '../../modules/music/infra/datasources/get_musics_by_artists_datasource.dart';
+import '../../modules/music/infra/repositories/get_musics_by_artist_repository_imp.dart';
 
 GetIt getIt = GetIt.I;
