--- lib/layers/domain/usecases/get_all_artists/get_all_artists_usecase.dart
@@ -1,6 +1,0 @@
-import 'package:dartz/dartz.dart';
-import 'package:music_app/layers/domain/entities/artist_entity.dart';
-
-abstract class GetAllArtistsUseCase {
-  Future<Either<Exception, List<ArtistEntity>>> getAllArtists();
-}
