--- lib/layers/domain/usecases/get_musics_by_artist/get_musics_by_artist_usecase.dart
@@ -1,8 +1,0 @@
-import 'package:dartz/dartz.dart';
-
-import '../../entities/music_entity.dart';
-
-abstract class GetMusicsByArtistUseCase {
-  Future<Either<Exception, List<MusicEntity>>> getMusicsByArtist(
-      String artistId);
-}
