+++ lib/modules/album/domain/usecases/get_albums_by_artist_usecase.dart
@@ -1,0 +1,8 @@
+import 'package:dartz/dartz.dart';
+
+import '../entities/album_entity.dart';
+
+abstract class GetAlbumsByArtistUseCase {
+  Future<Either<Exception, List<AlbumEntity>>> getAlbumsByArtist(
+      String artistId);
+}
