+++ lib/modules/artist/domain/repositories/get_all_artists_repository.dart
