+++ lib/modules/artist/domain/usecases/get_all_artists_usecase.dart
@@ -1,0 +1,7 @@
+import 'package:dartz/dartz.dart';
+
+import '../entities/artist_entity.dart';
+
+abstract class GetAllArtistsUseCase {
+  Future<Either<Exception, List<ArtistEntity>>> getAllArtists();
+}
