+++ lib/modules/music/domain/usecases/get_musics_by_artist_usecase.dart
@@ -1,0 +1,8 @@
+import 'package:dartz/dartz.dart';
+
+import '../entities/music_entity.dart';
+
+abstract class GetMusicsByArtistUseCase {
+  Future<Either<Exception, List<MusicEntity>>> getMusicsByArtist(
+      String artistId);
+}
