+++ lib/shared/enums/convertion_type_enum.dart
@@ -1,0 +1,5 @@
+enum ConversionType {
+  toUpperCase,
+  toLowerCase,
+  capitalize,
+}
