+++ pubspec.yaml
@@ -42,6 +42,7 @@ dependencies:
   dio: ^4.0.6
   bloc: ^8.1.0
   flutter_bloc: ^8.1.1
+  auto_size_text: ^3.0.0
 
 dev_dependencies:
   flutter_test:
